Move workflow version and sub-job retrieval to util
