+++ util/util.go
@@ -339,6 +339,34 @@ func GetWorkflowByID(id uuid.UUID) *types.Workflow {
 	return &workflow
 }
 
+func GetSubJobs(runID uuid.UUID) []types.SubJob {
+	if runID == uuid.Nil {
+		fmt.Println("Couldn't list sub-jobs, no run ID parameter specified!")
+		return nil
+	}
+	urlReq := "subjob/?execution=" + runID.String()
+	urlReq += "&page_size=" + strconv.Itoa(math.MaxInt)
+
+	resp := request.Trickest.Get().DoF(urlReq)
+	if resp == nil {
+		fmt.Println("Error: Couldn't get sub-jobs!")
+		return nil
+	}
+
+	if resp.Status() != http.StatusOK {
+		request.ProcessUnexpectedResponse(resp)
+	}
+
+	var subJobs types.SubJobs
+	err := json.Unmarshal(resp.Body(), &subJobs)
+	if err != nil {
+		fmt.Println("Error unmarshalling sub-jobs response!")
+		return nil
+	}
+
+	return subJobs.Results
+}
+
 func LabelSubJobs(subJobs []types.SubJob, version types.WorkflowVersionDetailed) []types.SubJob {
 	labels := make(map[string]bool)
 	for i := range subJobs {
@@ -376,6 +404,55 @@ func LabelSubJobs(subJobs []types.SubJob, version types.WorkflowVersionDetailed)
 		}
 	}
 	return subJobs
+}
+
+func GetWorkflowVersionByID(versionID, fleetID uuid.UUID) *types.WorkflowVersionDetailed {
+	resp := request.Trickest.Get().DoF("workflow-version/%s/", versionID)
+	if resp == nil {
+		fmt.Println("Error: Couldn't get workflow version!")
+		return nil
+	}
+
+	if resp.Status() != http.StatusOK {
+		request.ProcessUnexpectedResponse(resp)
+	}
+
+	var workflowVersion types.WorkflowVersionDetailed
+	err := json.Unmarshal(resp.Body(), &workflowVersion)
+	if err != nil {
+		fmt.Println("Error unmarshalling workflow version response!")
+		return nil
+	}
+
+	if fleetID != uuid.Nil {
+		maxMachines, err := GetWorkflowVersionMaxMachines(versionID.String(), fleetID)
+		if err != nil {
+			fmt.Printf("Error getting maximum machines: %v", err)
+			return nil
+		}
+		workflowVersion.MaxMachines = maxMachines
+
+	}
+	return &workflowVersion
+}
+
+func GetWorkflowVersionMaxMachines(version string, fleet uuid.UUID) (types.Machines, error) {
+	resp := request.Trickest.Get().DoF("workflow-version/%s/max-machines/?fleet=%s", version, fleet)
+	if resp == nil {
+		return types.Machines{}, fmt.Errorf("couldn't get workflow version's maximum machines")
+	}
+
+	if resp.Status() != http.StatusOK {
+		return types.Machines{}, fmt.Errorf("unexpected response status code for workflow version's maximum machines: %d", resp.Status())
+	}
+
+	var machines types.Machines
+	err := json.Unmarshal(resp.Body(), &machines)
+	if err != nil {
+		return types.Machines{}, fmt.Errorf("couldn't unmarshal workflow versions's maximum machines: %v", err)
+	}
+
+	return machines, nil
 }
 
 func ResolveObjectPath(path string, silent bool) (*types.SpaceDetailed, *types.Project, *types.Workflow, bool) {
